Fixed sorting Related to Issue [WTI-279]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@2002 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 19a4b047017471783a22a7088b0818ec99247aae

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/DuneAssessmentSectionEntity.cs
@@ -39,7 +39,7 @@ namespace Application.Ringtoets.Storage.DbContext
         public long ProjectEntityId { get; set; }
         public string Name { get; set; }
         public int Norm { get; set; }
-        public long Order { get; set; }
+        public int Order { get; set; }
     
         public virtual ProjectEntity ProjectEntity { get; set; }
     }
